T442: simplifying History trait
