+++ src/test/scala/scorex/consensus/mining/BlockGeneratorControllerSpecification.scala
@@ -35,7 +35,7 @@ class BlockGeneratorControllerSpecification extends ActorTestingCommons {
 
   private def historyWithLastBlock(block: Block, desiredHeight: Int): History = {
     val stubHistory = stub[History]
-    (stubHistory.lastBlock _).when().returns(block).anyNumberOfTimes()
+    (stubHistory.blockAt _).when(desiredHeight).returns(Some(block)).anyNumberOfTimes()
     (stubHistory.height _).when().returns(desiredHeight).anyNumberOfTimes()
     stubHistory
   }
